consumer/requirements.txt
pymongo[srv]
kafka-python
python-dotenv
